backend/requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
torch
transformers
numpy
pytesseract
pillow
